refactor(api.www.yantene.net): move value objects and entities to models directory
