+++ sites/api.www.yantene.net/src/domain/notes/models/body.value-object.ts
