+++ sites/api.www.yantene.net/src/domain/tags/tags.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { Tag } from "./entities/tag.entity";
+import { Tag } from "./models/tag.entity";
 
 @Injectable()
 export class TagsService {
